refactor(common): name the precision variant of interpolate_to_cell_center
`interpolate_to_cell_center` is the reduced-precision interpolation, but its
name did not say so, unlike the `_wp`/`_vp` suffixes used elsewhere. Renamed
to `interpolate_to_cell_center_vp` and its dycore callers updated.

Split out of C2SM/icon4py#1359, which is being separated into independently
reviewable pieces.
