+++ model/atmosphere/dycore/src/icon4py/model/atmosphere/dycore/stencils/compute_advection_in_horizontal_momentum_equation.py
@@ -21,8 +21,8 @@ from icon4py.model.atmosphere.dycore.stencils.mo_math_divrot_rot_vertex_ri_dsl i
 )
 from icon4py.model.common import dimension as dims, field_type_aliases as fa, type_alias as ta
 from icon4py.model.common.dimension import E2C, E2C2EO, E2V
-from icon4py.model.common.interpolation.stencils.interpolate_to_cell_center import (
-    _interpolate_to_cell_center,
+from icon4py.model.common.interpolation.stencils.interpolate_to_cell_center_vp import (
+    _interpolate_to_cell_center_vp,
 )
 from icon4py.model.common.type_alias import vpfloat, wpfloat
 
@@ -41,7 +41,7 @@ def _compute_advective_normal_wind_tendency(
     ddqz_z_full_e: fa.EdgeKField[ta.vpfloat],
 ) -> fa.EdgeKField[ta.vpfloat]:
     #: intermediate variable horizontal_kinetic_energy_at_cells_on_model_levels is originally declared as z_ekinh in ICON
-    horizontal_kinetic_energy_at_cells_on_model_levels = _interpolate_to_cell_center(
+    horizontal_kinetic_energy_at_cells_on_model_levels = _interpolate_to_cell_center_vp(
         horizontal_kinetic_energy_at_edges_on_model_levels, e_bln_c_s
     )
     horizontal_kinetic_energy_at_cells_on_model_levels = astype(
