+++ model/common/src/icon4py/model/common/interpolation/stencils/interpolate_to_cell_center_vp.py
@@ -14,18 +14,22 @@ from icon4py.model.common.type_alias import vpfloat, wpfloat
 
 
 @gtx.field_operator
-def _interpolate_to_cell_center(
+def _interpolate_to_cell_center_vp(
     interpolant: fa.EdgeKField[vpfloat],
     e_bln_c_s: gtx.Field[gtx.Dims[dims.CellDim, dims.C2EDim], wpfloat],
 ) -> fa.CellKField[vpfloat]:
-    """Formerly known as mo_velocity_advection_stencil_08 or mo_velocity_advection_stencil_09."""
+    """Interpolate an edge field to the cell centers with the bilinear C2E weights.
+
+    Formerly known as mo_velocity_advection_stencil_08 or
+    mo_velocity_advection_stencil_09.
+    """
     interpolant_wp = astype(interpolant, wpfloat)
     interpolation_wp = neighbor_sum(e_bln_c_s * interpolant_wp(C2E), axis=dims.C2EDim)
     return astype(interpolation_wp, vpfloat)
 
 
 @gtx.program(grid_type=gtx.GridType.UNSTRUCTURED)
-def interpolate_to_cell_center(
+def interpolate_to_cell_center_vp(
     interpolant: fa.EdgeKField[vpfloat],
     e_bln_c_s: gtx.Field[gtx.Dims[dims.CellDim, dims.C2EDim], wpfloat],
     interpolation: fa.CellKField[vpfloat],
@@ -34,7 +38,7 @@ def interpolate_to_cell_center(
     vertical_start: gtx.int32,
     vertical_end: gtx.int32,
 ) -> None:
-    _interpolate_to_cell_center(
+    _interpolate_to_cell_center_vp(
         interpolant=interpolant,
         e_bln_c_s=e_bln_c_s,
         out=interpolation,
